Move mapping function from FollowSetFeedFilter to FollowSet.
