+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/dal/FollowSetFeedFilter.kt
@@ -22,38 +22,11 @@ package com.vitorpamplona.amethyst.ui.dal
 
 import com.vitorpamplona.amethyst.model.Account
 import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.quartz.events.PeopleListEvent
 
 class FollowSetFeedFilter(
     val account: Account,
 ) : FeedFilter<Note>() {
     override fun feedKey(): String = account.userProfile().pubkeyHex
-
-    private fun mapEventToSet(event: PeopleListEvent): PeopleListEvent.FollowSet {
-        val dTag = event.dTag()
-        val listTitle = event.title() ?: dTag
-        val listDescription = event.description() ?: ""
-        val publicFollows = event.filterTagList("p", null)
-        val privateFollows = mutableListOf<String>()
-        event.privateTaggedUsers(account.signer) { userList -> privateFollows.addAll(userList) }
-        return if (publicFollows.isEmpty() && privateFollows.isNotEmpty()) {
-            PeopleListEvent.FollowSet(
-                isPrivate = true,
-                title = listTitle,
-                description = listDescription,
-                profileList = privateFollows.toSet(),
-            )
-        } else if (publicFollows.isNotEmpty() && privateFollows.isEmpty()) {
-            PeopleListEvent.FollowSet(
-                isPrivate = false,
-                title = listTitle,
-                description = listDescription,
-                profileList = publicFollows.toSet(),
-            )
-        } else {
-            throw Exception("Mixed follow sets are not yet supported.")
-        }
-    }
 
     override fun feed(): List<Note> =
         account
